+++ CopperDevs.Windowing.SDL3/SDL3Window.cs
@@ -1,5 +1,6 @@
 using CopperDevs.Core.Data;
 using CopperDevs.Core.Utility;
+using CopperDevs.Logger;
 using CopperDevs.Windowing.Data;
 using CopperDevs.Windowing.SDL3.Data;
 
@@ -95,6 +96,25 @@ public class SDL3Window : Window
 
         SDL.SetHint(SDL_HINT_WINDOWS_CLOSE_ON_ALT_F4, "1"u8);
         SDL.SetHint(SDL_HINT_WINDOWS_CLOSE_ON_ALT_F4, "1");
+
+        {
+            SetProperty(AppMetadata.MetadataProperty.Name, options.Metadata);
+            SetProperty(AppMetadata.MetadataProperty.Version, options.Metadata);
+            SetProperty(AppMetadata.MetadataProperty.Identifier, options.Metadata);
+            SetProperty(AppMetadata.MetadataProperty.Creator, options.Metadata);
+            SetProperty(AppMetadata.MetadataProperty.Copyright, options.Metadata);
+            SetProperty(AppMetadata.MetadataProperty.Url, options.Metadata);
+            SetProperty(AppMetadata.MetadataProperty.Type, options.Metadata);
+
+            void SetProperty(AppMetadata.MetadataProperty property, AppMetadata metadata)
+            {
+                Log.Config($"Setting metadata property {property.ToString()} to '{metadata.GetProperty(property)}'");
+                if (SDL.SetAppMetadataProperty(property, metadata.GetProperty(property)))
+                    Log.Success($"Metadata property {property.ToString()} changed to '{SDL.GetAppMetadataProperty(property)}'");
+                else
+                    Log.Error($"Failed to set metadata property {property.ToString()} to {metadata.GetProperty(property)}. Error: {SDL.GetError()}");
+            }
+        }
 
         window = new ManagedSDLWindow(options);
         window.HandleEvent += EventsHandler;
